Fix tests for membership
1. Moved tests to `test_util`
2. Changed them so that they work with new `get_members` required
   arguments
